Resolving merge conflict
